chore: move `IncomingOrganizationManager` to `managers.py`
